返回更新行数

+++ src/main/java/com/d1m/wechat/domain/dao/MemberTagDataDao.java
@@ -10,6 +10,6 @@ import java.util.Collection;
  */
 public interface MemberTagDataDao {
 
-    long updateBatch(Collection<MemberTagData> records);
+    int updateBatch(Collection<MemberTagData> records);
 
 }
